refactor(popover): rename scss files to use .module.scss
