+++ draft-packages/popover/KaizenDraft/Popover/classMappers.ts
@@ -1,9 +1,9 @@
-import * as React from "react"
+import React from "react"
 import informativeIcon from "@kaizen/component-library/icons/information.icon.svg"
 import positiveIcon from "@kaizen/component-library/icons/success.icon.svg"
 import negativeIcon from "@kaizen/component-library/icons/exclamation.icon.svg"
-import styles from "./styles.scss"
 import { Size, Variant } from "./types"
+import styles from "./Popover.module.scss"
 
 export const mapVariantToBoxClass = (variant: Variant): string => {
   switch (variant) {
@@ -76,10 +76,5 @@ export const mapSizeToClass = (size: Size): string => {
   }
 }
 
-export const mapLineVariant = (singleLine: boolean): string => {
-  if (singleLine === true) {
-    return styles.singleLine
-  } else {
-    return ""
-  }
-}
+export const mapLineVariant = (singleLine: boolean): string =>
+  singleLine ? styles.singleLine : ""
